Partially fixed problem with enums without definitions/projects
